Fixed a test per Kens review.  Consolidated trajectory parameter tests into a single file.
